Enhance code quality
* Group DirWatcher and Importer modules into a special directory
* Import only promisify function from utility module
