+++ app.js
@@ -2,9 +2,7 @@ import path from 'path';
 
 import config from './config';
 import { Product, User } from './models';
-
-import DirWatcher from './dirwatcher';
-import Importer from './importer';
+import { DirWatcher, Importer } from './modules';
 
 const dirWatcher = new DirWatcher();
 const importer = new Importer();
@@ -13,13 +11,14 @@ const delay = config.watcherDelay;
 
 dirWatcher.on('dirwatcher:changed', async path => {
   const json = await importer.import(path);
-  console.log(`${path} has been changed`);
+  console.log(`Async ${path} has been changed`);
   console.log(json);
 });
 
 dirWatcher.on('dirwatcher:changed', path => {
+  const json = importer.importSync(path);
   console.log(`Sync ${path} has been changed`);
-  console.log(importer.importSync(path));
+  console.log(json);
 });
 
 dirWatcher.on('dirwatcher:deleted', path => {
